\name{SurvLong-package}
\alias{SurvLong-package}
\alias{SurvLong}
\docType{package}
\title{
Analysis of Proportional Hazards Model with Sparse Longitudinal Covariates
}
\description{
Kernel weighting methods for estimation of proportional hazards 
models with intermittently observed longitudinal covariates.}
\details{
\tabular{ll}{
Package: \tab SurvLong\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-22\cr
License: \tab GPL-2\cr
}
}
\author{
Hongyuan Cao, Mathew M. Churpek, Donglin Zeng, Jason P. Fine, and Shannon T. Holloway

Maintainer: Shannon T. Holloway <sthollow@ncsu.edu>
}
\references{
Cao, H., Churpek, M. M., Zeng, D., and Fine, J. P. (2015).
Journal of the American Statistical Association, in press.
}
\keyword{ package }
\seealso{
         \code{\link{fullKernel}}, 
         \code{\link{halfKernel}}, 
         \code{\link{lastValue}}, 
         \code{\link{nearValue}}
}

