\name{SurvLong-package}
\alias{SurvLong-package}
\alias{SurvLong}
\docType{package}
\title{
Analysis of Proportional Hazards Model with Sparse Longitudinal Covariates
}
\description{
Kernel weighting methods for estimation of proportional hazards 
models with intermittently observed longitudinal covariates. Cao H., Churpek M. M., Zeng D., and Fine J. P. (2015) <doi:10.1080/01621459.2014.957289>}
\details{
\tabular{ll}{
Package: \tab SurvLong\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2020-01-08\cr
License: \tab GPL-2\cr
}
}
\author{
Hongyuan Cao, Mathew M. Churpek, Donglin Zeng, Jason P. Fine, and Shannon T. Holloway

Maintainer: Shannon T. Holloway <sthollow@ncsu.edu>
}
\references{
Cao H., Churpek M. M., Zeng D., Fine J. P. 
(2015).
Analysis of the proportional hazards model with sparse longitudinal covariates.
Journal of the American Statistical Association, 110, 1187-1196.
}
\keyword{ package }
\seealso{
         \code{\link{fullKernel}}, 
         \code{\link{halfKernel}}, 
         \code{\link{lastValue}}, 
         \code{\link{nearValue}}
}

